Transfer of the price dropbox in the preferences pane. Moving of the modules list compiling from priceview to price service.
